test: move webhook signature checks out of generated tests
